Upload content migrated + refactored utils folder.
