--- backend/src/controllers/mod.rs
@@ -1,1 +1,0 @@
-pub mod upload_controller;
