+++ backend/src/utils/ip.rs
@@ -3,6 +3,7 @@ use actix_web::HttpRequest;
 pub struct IPUtils;
 
 impl IPUtils {
+    // Unused?
     pub fn get_ip_from_request(req: &HttpRequest) -> Option<String> {
         if let Some(ip) = req.headers().get("Cf-Connecting-IP").and_then(|v| v.to_str().ok()) {
             return Some(Self::extract_first_ip(ip));
@@ -15,7 +16,7 @@ impl IPUtils {
         req.peer_addr()
             .map(|addr| addr.ip().to_string())
     }
-
+    // Also unused??
     fn extract_first_ip(ip_list: &str) -> String {
         ip_list.split(',').next().unwrap_or("").trim().to_string()
     }
