+++ backend/src/utils/mod.rs
@@ -1,2 +1,3 @@
-pub mod iputils;
-pub mod datautils;
+pub mod ip;
+pub mod data;
+pub mod string;
